+++ qml/pages/ChatSettingsPage.qml
@@ -43,11 +43,12 @@ Page {
     function init () {
 
         // Get the member status of the user himself
-        storage.transaction ( "SELECT description, membership, power_event_name, power_kick, power_ban, power_invite, power_event_power_levels, power_event_avatar FROM Chats WHERE id='" + activeChat + "'", function (res) {
+        storage.transaction ( "SELECT description, avatar_url, membership, power_event_name, power_kick, power_ban, power_invite, power_event_power_levels, power_event_avatar FROM Chats WHERE id='" + activeChat + "'", function (res) {
 
             description = res.rows[0].description
             storage.transaction ( "SELECT * FROM Memberships WHERE chat_id='" + activeChat + "' AND matrix_id='" + matrix.matrixid + "'", function (membershipResult) {
                 membership = membershipResult.rows[0].membership
+                avatarImage.mxc = membershipResult.rows[0].avatar_url
                 power = membershipResult.rows[0].power_level
                 canChangeName = power >= res.rows[0].power_event_name
                 canKick = power >= res.rows[0].power_kick
@@ -55,13 +56,10 @@ Page {
                 canInvite = power >= res.rows[0].power_invite
                 canChangeAvatar = power >= res.rows[0].power_event_avatar
                 canChangePermissions = power >= res.rows[0].power_event_power_levels
-
+                console.log("AVATARURL:", membershipResult.rows[0].avatar_url)
                 console.log("POWER:", power, "canChangeAvatar:", res.rows[0].power_event_avatar, JSON.stringify(res.rows[0]))
             })
         })
-
-        // Get the chat avatar
-        roomnames.getAvatarUrl ( activeChat, function ( avatar_url ) { avatarImage.mxc = avatar_url } )
 
         // Request the full memberlist, from the database
         model.clear()
@@ -151,6 +149,7 @@ Page {
                 width: parent.width / 2
                 anchors.horizontalCenter: parent.horizontalCenter
                 mxc: ""
+                visible: mxc !== "" && mxc !== undefined
                 onClickFunction: function () {
                     var hasAvatar = avatarImage.mxc !== "" && avatarImage.mxc !== null
                     if ( canChangeAvatar && hasAvatar ) contextualAvatarActions.show()